@@ -19,13 +19,13 @@
   "author": "Carlos Lapao",
   "license": "MIT",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.29.2",
-    "@typescript-eslint/parser": "^4.29.2",
-    "electron": "^13.2.1",
-    "eslint": "^7.32.0",
-    "typescript": "^4.3.5",
     "@electron/remote": "1.2.0",
     "@types/node": "16.9.4",
-    "axios": "0.21.4"
+    "@typescript-eslint/eslint-plugin": "^4.29.2",
+    "@typescript-eslint/parser": "^4.29.2",
+    "axios": "0.21.4",
+    "electron": "^13.4.0",
+    "eslint": "^7.32.0",
+    "typescript": "^4.3.5"
   }
 }
