@@ -1,14 +1,32 @@
+import { SettingsComponent } from './components/settings/settings.component';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
+import { SideMenuComponent } from './components/side-menu/side-menu.component';
+import { StatusBarComponent } from './components/status-bar/status-bar.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
-  declarations: [AppComponent],
-  imports: [BrowserModule, HttpClientModule, AppRoutingModule],
+  declarations: [
+    AppComponent,
+    SideMenuComponent,
+    StatusBarComponent,
+    SettingsComponent,
+  ],
+  imports: [
+    BrowserModule,
+    FormsModule,
+    ReactiveFormsModule,
+    HttpClientModule,
+    AppRoutingModule,
+    FontAwesomeModule,
+  ],
   providers: [],
+  exports: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
