@@ -1,8 +1,19 @@
+import { HomeComponent } from './components/home/home.component';
+import { SettingsComponent } from './components/settings/settings.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 // import { MinikubeService } from './services/minikube.service';
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: 'settings',
+    component: SettingsComponent,
+  },
+  {
+    path: '**',
+    component: HomeComponent,
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
