@@ -1,7 +1,10 @@
-import { DevelopmentTool } from './entities/tool';
-import { KubectlService } from './services/kubectl.service';
+import {
+  DevelopmentTool,
+  DevelopmentToolStatus,
+} from './../../app/src/interfaces/development-tool';
 import { Component } from '@angular/core';
 import { MinikubeService } from './services/minikube.service';
+import { BaseDevelopmentToolService } from './services/base-development-tool.service';
 
 @Component({
   selector: 'app-root',
@@ -10,36 +13,4 @@ import { MinikubeService } from './services/minikube.service';
 })
 export class AppComponent {
   title = 'electron-angular-helloworld';
-  public status = 'unknown';
-  public kubectlVersion = 'unknown';
-  public tools: DevelopmentTool[];
-
-  constructor(
-    private minikube: MinikubeService,
-    private kubectl: KubectlService
-  ) {
-    this.tools = [];
-    this.init();
-  }
-
-  async init(): Promise<void> {
-    const kubecltTool = await this.kubectl.init();
-    const minikubeTool = await this.minikube.init();
-    this.tools.push(kubecltTool);
-    this.tools.push(minikubeTool);
-  }
-
-  async test() {
-    let t = await this.minikube.test();
-    console.log(t);
-  }
-
-  async testParameters(): Promise<void> {}
-
-  async getMinikubeStatus() {
-    this.status = 'reading';
-    this.status = (await this.minikube.getMinikubeStatus()).toString();
-    await this.kubectl.init();
-    console.log('testing');
-  }
 }
